refactor(preact/eslint): moving eslint out of toolbox and adjusting the package files
