+++ packages/snap-preact/eslint/jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
 	displayName: 'eslint-plugin-snap-config',
 	testEnvironment: 'node',
-	testMatch: ['<rootDir>/*.test.cjs'],
+	testMatch: ['<rootDir>/tests/*.test.cjs'],
 	transform: {},
 };
