+++ packages/snap-preact/eslint/src/index.cjs
