+++ packages/snap-preact/eslint/src/validate-config.cjs
